+++ python/tests/test_vcf.py
@@ -200,7 +200,9 @@ class ExamplesMixin:
 
     def test_simple_infinite_sites_random_ploidy(self):
         ts = msprime.simulate(10, mutation_rate=1, random_seed=2)
-        ts = tsutil.insert_random_ploidy_individuals(ts, min_ploidy=1)
+        ts = tsutil.insert_random_ploidy_individuals(
+            ts, min_ploidy=1, samples_only=True
+        )
         assert ts.num_sites > 2
         self.verify(ts)
 
@@ -227,7 +229,9 @@ class ExamplesMixin:
     def test_simple_jukes_cantor_random_ploidy(self):
         ts = msprime.simulate(10, random_seed=2)
         ts = tsutil.jukes_cantor(ts, num_sites=10, mu=1, seed=2)
-        ts = tsutil.insert_random_ploidy_individuals(ts, min_ploidy=1)
+        ts = tsutil.insert_random_ploidy_individuals(
+            ts, min_ploidy=1, samples_only=True
+        )
         self.verify(ts)
 
     def test_single_tree_multichar_mutations(self):